Move KernelKey and clean_up into their own infra module
