+++ tests/blackbox/stratis_cli_cert.py
@@ -23,7 +23,7 @@ import time
 import unittest
 
 # isort: THIRDPARTY
-from testlib.dbus import KernelKey, clean_up
+from testlib.infra import KernelKey, clean_up
 from testlib.stratis import STRATIS_CLI
 from testlib.utils import (
     RandomKeyTmpFile,
